+++ packages/snap-preact-demo/src/components/Recommendations/Recs/Recs.tsx
@@ -35,7 +35,7 @@ export class Recs extends Component<RecsProps> {
 				<hr style={{ margin: '20px 0' }} />
 				<Recommendation controller={controller} title={'Recommended For You'} speed={0}>
 					{store.results.map((result) => (
-						<Result result={result}></Result>
+						<Result controller={controller} result={result}></Result>
 					))}
 				</Recommendation>
 			</div>